pie chart typescript conversion
