--- src/components/PieChart/PieChart.reducers.js
@@ -1,14 +1,0 @@
-export function onMouseOver(state, hoveringIndex) {
-	return {
-		...state,
-		isHovering: true,
-		hoveringIndex,
-	};
-}
-
-export function onMouseOut(state) {
-	return {
-		...state,
-		isHovering: false,
-	};
-}
