+++ src/components/PieChart/PieChart.reducers.ts
@@ -1,0 +1,20 @@
+export interface IPieChartState {
+	isHovering: boolean,
+	hoveringIndex: number
+}
+
+export default {
+	onMouseOut(state: IPieChartState) {
+		return {
+			...state,
+			isHovering: false,
+		};
+	},
+	onMouseOver(state: IPieChartState, hoveringIndex: number) {
+		return {
+			...state,
+			isHovering: true,
+			hoveringIndex,
+		};
+	}
+}
